+++ NotificationServices/src/main/java/com/example/notificationservices/notification/NotificationsRepository.java
@@ -3,7 +3,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface NotificationsRepository extends JpaRepository<Notification, Long>{
+public interface NotificationsRepository extends JpaRepository<Notification, Integer>{
 
 
 }
